Finished module Basics_part_II

+++ src/main/java/Util/Prime.java
@@ -6,7 +6,7 @@ public class Prime {
         if (number <= 1) return false;
         else if (number == 2 || number == 3) return true;
         else if (number % 2 == 0 || number % 3 == 0) return false;
-        for (long i = 5; i <= Math.sqrt(number); i += 6) {
+        for (int i = 5; i <= Math.sqrt(number); i += 6) {
             if (number % i == 0 || number % (i+2) == 0) return false;
         }
         return true;
